.ipynb_checkpoints/requirement-checkpoint.txt
pyyaml
tqdm
munch
resampy
soundfile
moviepy==1.0.3
gast==0.2.2
ipython
jupyter
matplotlib
pandas
xlrd
openpyxl
tomorrow3
opencv-python
tensorflow-gpu==1.14
torch==1.4
torchvision==0.5.0
ml_collections
tqdm
efficientnet_pytorch
transformers
catalyst
sklearn
opencv-python
albumentations